+++ src/redis/src/Redis.php
@@ -45,8 +45,8 @@ class Redis
             // Release connection.
             if (! $hasContextConnection) {
                 if ($this->shouldUseSameConnection($name)) {
-                    if ($name == 'select') {
-                        $connection->setDbChanged(true);
+                    if ($name === 'select' && $db = $arguments[0]) {
+                        $connection->setDatabase((int) $db);
                     }
                     // Should storage the connection to coroutine context, then use defer() to release the connection.
                     Context::set($this->getContextKey(), $connection);