+++ src/redis/tests/RedisTest.php
@@ -19,7 +19,6 @@ use Hyperf\Redis\Pool\PoolFactory;
 use Hyperf\Redis\Pool\RedisPool;
 use Hyperf\Redis\Redis;
 use Hyperf\Utils\ApplicationContext;
-use Hyperf\Utils\Context;
 use HyperfTest\Redis\Stub\RedisPoolStub;
 use Mockery;
 use PHPUnit\Framework\TestCase;
@@ -49,12 +48,45 @@ class RedisTest extends TestCase
         $this->assertTrue($redis->connect('127.0.0.1', 6379, 0.0));
     }
 
-    public function testRedisCommand()
+    public function testRedisSelect()
+    {
+        $redis = $this->getRedis();
+
+        $res = $redis->set('xxxx', 'yyyy');
+        $this->assertSame('db:0 name:set argument:xxxx,yyyy', $res);
+
+        $redis->select(2);
+        $res = $redis->get('xxxx');
+        $this->assertSame('db:2 name:get argument:xxxx', $res);
+
+        $this->assertSame(2, $redis->getDatabase());
+
+        $res = parallel([function () use ($redis) {
+            return $redis->get('xxxx');
+        }]);
+
+        $this->assertSame('db:0 name:get argument:xxxx', $res[0]);
+    }
+
+    private function getRedis()
     {
         $container = Mockery::mock(Container::class);
         $container->shouldReceive('get')->once()->with(ConfigInterface::class)->andReturn(new Config([
             'redis' => [
-                'default' => [],
+                'default' => [
+                    'host' => 'localhost',
+                    'auth' => null,
+                    'port' => 6379,
+                    'db' => 0,
+                    'pool' => [
+                        'min_connections' => 1,
+                        'max_connections' => 30,
+                        'connect_timeout' => 10.0,
+                        'wait_timeout' => 3.0,
+                        'heartbeat' => -1,
+                        'max_idle_time' => 60,
+                    ],
+                ],
             ],
         ]));
         $pool = new RedisPoolStub($container, 'default');
@@ -64,18 +96,6 @@ class RedisTest extends TestCase
 
         $factory = new PoolFactory($container);
 
-        $redis = new Redis($factory);
-
-        $res = $redis->set('xxxx', 'yyyy');
-        $this->assertSame('db:0 name:set argument:xxxx,yyyy', $res);
-
-        $redis->select(2);
-        $res = $redis->get('xxxx');
-        $this->assertSame('db:2 name:get argument:xxxx', $res);
-
-        parallel([function () use ($redis) {
-            $res = $redis->get('xxxx');
-            $this->assertSame('db:0 name:get argument:xxxx', $res);
-        }]);
+        return new Redis($factory);
     }
 }